fix: make progress logic shared
